Add CopyInput component to @valist/ui
