+++ packages/valist-ui/src/components/CopyInput/index.ts
@@ -1,0 +1,1 @@
+export * from './CopyInput';
