+++ packages/valist-web/components/DeployForm/screens/AddKey.tsx
@@ -1,4 +1,5 @@
 import { Text, TextInput, Title } from "@mantine/core";
+import { CopyInput } from "@valist/ui";
 
 interface AddKeyProps {
   account: string;
@@ -16,11 +17,9 @@ export function AddKey(props: AddKeyProps):JSX.Element {
       <Title order={4}>A new Ethereum key has been securely generated by your browser:</Title>
       <br />
       <div>
-        <TextInput
-          disabled
-          value={props.publicKey} 
+        <CopyInput
+          value={props.publicKey}
         />
-
         <br />
         <Text>1. The new <b>public key address</b> will be added to the <b>{props.account}/{props.project}</b> Valist project.</Text>
         <Text>2. The new <b>private key</b> will be added to the <b>{props.repo}</b> GitHub repo as a secret named <b>VALIST_SIGNER</b>.</Text>
